Move setup_anvil to test_helpers
